admin/app/requirements/base.txt
django==4.2.11
python-dotenv==1.0.1
psycopg2-binary==2.9.9
django-debug-toolbar==4.3.0
django-split-settings==1.3.1
gunicorn==22.0.0
django-environ==0.11.2
djangorestframework==3.15.1
django-constance==3.1.0
django-redis==5.4.0
drf-spectacular==0.27.2
